src/analytics-ai/event-tracking-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
aiokafka==0.10.0
pydantic==2.7.1
pydantic-settings==2.2.1
cassandra-driver==3.29.1
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
